과제 생성 api(피로체크 전용)

+++ backend/src/main/java/com/example/Piroin/project/domain/assignment/repository/AssignmentItemRepository.java
@@ -3,5 +3,5 @@ package com.example.Piroin.project.domain.assignment.repository;
 import com.example.Piroin.project.domain.assignment.entity.AssignmentItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssignmentItemRepository extends JpaRepository<AssignmentItem, Long> {
+public interface AssignmentItemRepository extends JpaRepository<AssignmentItem, Integer> {
 }
